// frontend/src/api/models/repairResponseStato.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Gestionale Negozio API
 * API del gestionale negozio
 * OpenAPI spec version: 1.0.0
 */

export type RepairResponseStato = typeof RepairResponseStato[keyof typeof RepairResponseStato];


export const RepairResponseStato = {
  NUOVO: 'NUOVO',
  IN_CORSO: 'IN_CORSO',
  PRONTO: 'PRONTO',
  CONSEGNATO: 'CONSEGNATO',
} as const;